requirements.txt
Werkzeug==0.14.1
pymongo==3.7.2
Flask==1.0.2
flask_restplus==0.12.1
PyJWT==1.7.1
